requirements.txt
pdfplumber
pytesseract
Pillow
streamlit
groq
numpy
requests
faiss-cpu; python_version < "3.13"
